Fixed RichAlignmentEventSelection.cpp based on Coverity
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@147555 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: ccd0fce4d513d478e4be751f89c89b668f4659f6

+++ Rec/Rich/RichAlignment/src/RichAlignmentEventSelection.cpp
@@ -140,7 +140,7 @@ StatusCode EventSelection::execute()
       if ( !unAmbiguousPhoton ) { continue; }
 
       // mirror combination number
-      const unsigned long long key = (100000*sphMirNum) + flatMirNum;
+      const unsigned long long key = (100000*(unsigned long long)sphMirNum) + flatMirNum;
 
       // count the photons in each of the mirror combinations
       ++m_combiMap[key];
